+++ packages/stacked_tools/lib/src/templates/compiled_templates.dart
@@ -1,6 +1,7 @@
 /// NOTE: This is generated code from the compileTemplates command. Do not modify by hand
 ///       This file should be checked into source control.
 
+
 // -------- HomeViewmodelTest Template Data ----------
 
 const String kAppTemplateHomeViewmodelTestPath =
@@ -52,9 +53,11 @@ void main() {
 
 // --------------------------------------------------
 
+
 // -------- TestHelpers Template Data ----------
 
-const String kAppTemplateTestHelpersPath = 'test/helpers/test_helpers.dart.stk';
+const String kAppTemplateTestHelpersPath =
+    'test/helpers/test_helpers.dart.stk';
 
 const String kAppTemplateTestHelpersContent = '''
 import 'package:mockito/annotations.dart';
@@ -140,9 +143,11 @@ void _removeRegistrationIfExists<T extends Object>() {
 
 // --------------------------------------------------
 
+
 // -------- StackedConfigJsonStk Template Data ----------
 
-const String kAppTemplateStackedConfigJsonStkPath = 'stacked.config.json.stk';
+const String kAppTemplateStackedConfigJsonStkPath =
+    'stacked.config.json.stk';
 
 const String kAppTemplateStackedConfigJsonStkContent = '''
 {
@@ -161,9 +166,11 @@ const String kAppTemplateStackedConfigJsonStkContent = '''
 
 // --------------------------------------------------
 
+
 // -------- READMEMdStk Template Data ----------
 
-const String kAppTemplateREADMEMdStkPath = 'README.md.stk';
+const String kAppTemplateREADMEMdStkPath =
+    'README.md.stk';
 
 const String kAppTemplateREADMEMdStkContent = '''
 # stacked_app
@@ -187,9 +194,11 @@ samples, guidance on mobile development, and a full API reference.
 
 // --------------------------------------------------
 
+
 // -------- Main Template Data ----------
 
-const String kAppTemplateMainPath = 'lib/main.dart.stk';
+const String kAppTemplateMainPath =
+    'lib/main.dart.stk';
 
 const String kAppTemplateMainContent = '''
 import 'package:flutter/material.dart';
@@ -234,6 +243,7 @@ class MyApp extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- SetupDialogUi Template Data ----------
 
 const String kAppTemplateSetupDialogUiPath =
@@ -259,6 +269,7 @@ void setupDialogUi() {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- SetupBottomSheetUi Template Data ----------
 
@@ -290,9 +301,11 @@ void setupBottomSheetUi() {
 
 // --------------------------------------------------
 
+
 // -------- UiHelpers Template Data ----------
 
-const String kAppTemplateUiHelpersPath = 'lib/ui/common/ui_helpers.dart.stk';
+const String kAppTemplateUiHelpersPath =
+    'lib/ui/common/ui_helpers.dart.stk';
 
 const String kAppTemplateUiHelpersContent = '''
 import 'dart:math';
@@ -378,9 +391,11 @@ double getResponsiveFontSize(BuildContext context,
 
 // --------------------------------------------------
 
+
 // -------- AppStrings Template Data ----------
 
-const String kAppTemplateAppStringsPath = 'lib/ui/common/app_strings.dart.stk';
+const String kAppTemplateAppStringsPath =
+    'lib/ui/common/app_strings.dart.stk';
 
 const String kAppTemplateAppStringsContent = '''
 const String ksHomeBottomSheetTitle = 'Build Great Apps!';
@@ -391,9 +406,11 @@ const String ksHomeBottomSheetDescription =
 
 // --------------------------------------------------
 
+
 // -------- AppColors Template Data ----------
 
-const String kAppTemplateAppColorsPath = 'lib/ui/common/app_colors.dart.stk';
+const String kAppTemplateAppColorsPath =
+    'lib/ui/common/app_colors.dart.stk';
 
 const String kAppTemplateAppColorsContent = '''
 import 'package:flutter/material.dart';
@@ -409,6 +426,7 @@ const Color kcBackgroundColor = kcDarkGreyColor;
 ''';
 
 // --------------------------------------------------
+
 
 // -------- NoticeSheet Template Data ----------
 
@@ -467,6 +485,7 @@ class NoticeSheet extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- NoticeSheetViewmodel Template Data ----------
 
 const String kAppTemplateNoticeSheetViewmodelPath =
@@ -477,6 +496,7 @@ const String kAppTemplateNoticeSheetViewmodelContent = '''
 ''';
 
 // --------------------------------------------------
+
 
 // -------- InfoAlertDialog Template Data ----------
 
@@ -586,6 +606,7 @@ class InfoAlertDialog extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- InfoAlertViewmodel Template Data ----------
 
 const String kAppTemplateInfoAlertViewmodelPath =
@@ -598,6 +619,7 @@ const String kAppTemplateInfoAlertViewmodelContent = '''
 ''';
 
 // --------------------------------------------------
+
 
 // -------- HomeViewV1 Template Data ----------
 
@@ -687,9 +709,11 @@ class HomeView extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- HomeView Template Data ----------
 
-const String kAppTemplateHomeViewPath = 'lib/ui/views/home/home_view.dart.stk';
+const String kAppTemplateHomeViewPath =
+    'lib/ui/views/home/home_view.dart.stk';
 
 const String kAppTemplateHomeViewContent = '''
 import 'package:flutter/material.dart';
@@ -703,7 +727,11 @@ class HomeView extends StackedView<HomeViewModel> {
   const HomeView({Key? key}) : super(key: key);
 
   @override
-  Widget builder(BuildContext context, HomeViewModel model, Widget? child) {
+  Widget builder(
+    BuildContext context,
+    HomeViewModel viewModel,
+    Widget? child,
+  ) {
     return Scaffold(
       body: SafeArea(
         child: Padding(
@@ -726,9 +754,9 @@ class HomeView extends StackedView<HomeViewModel> {
                     verticalSpaceMedium,
                     MaterialButton(
                       color: Colors.black,
-                      onPressed: model.incrementCounter,
+                      onPressed: viewModel.incrementCounter,
                       child: Text(
-                        model.counterLabel,
+                        viewModel.counterLabel,
                         style: const TextStyle(color: Colors.white),
                       ),
                     ),
@@ -745,7 +773,7 @@ class HomeView extends StackedView<HomeViewModel> {
                           color: Colors.white,
                         ),
                       ),
-                      onPressed: model.showDialog,
+                      onPressed: viewModel.showDialog,
                     ),
                     MaterialButton(
                       color: kcDarkGreyColor,
@@ -755,7 +783,7 @@ class HomeView extends StackedView<HomeViewModel> {
                           color: Colors.white,
                         ),
                       ),
-                      onPressed: model.showBottomSheet,
+                      onPressed: viewModel.showBottomSheet,
                     ),
                   ],
                 )
@@ -768,14 +796,15 @@ class HomeView extends StackedView<HomeViewModel> {
   }
 
   @override
-  HomeViewModel modelBuilder(BuildContext context) {
-    return HomeViewModel();
-  }
+  HomeViewModel viewModelBuilder(
+    BuildContext context,
+  ) => HomeViewModel();
 }
 
 ''';
 
 // --------------------------------------------------
+
 
 // -------- HomeViewmodel Template Data ----------
 
@@ -823,6 +852,7 @@ class HomeViewModel extends BaseViewModel {
 
 // --------------------------------------------------
 
+
 // -------- StartupViewmodel Template Data ----------
 
 const String kAppTemplateStartupViewmodelPath =
@@ -851,6 +881,7 @@ class StartupViewModel extends BaseViewModel {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- StartupViewV1 Template Data ----------
 
@@ -917,6 +948,7 @@ class StartupView extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- StartupView Template Data ----------
 
 const String kAppTemplateStartupViewPath =
@@ -934,7 +966,11 @@ class StartupView extends StackedView<StartupViewModel> {
   const StartupView({Key? key}) : super(key: key);
 
   @override
-  Widget builder(BuildContext context, StartupViewModel model, Widget? child) {
+  Widget builder(
+    BuildContext context,
+    StartupViewModel viewModel,
+    Widget? child,
+  ) {
     return Scaffold(
       body: Center(
         child: Column(
@@ -974,22 +1010,24 @@ class StartupView extends StackedView<StartupViewModel> {
   }
 
   @override
-  StartupViewModel modelBuilder(BuildContext context) {
-    return StartupViewModel();
-  }
+  StartupViewModel viewModelBuilder(
+    BuildContext context,
+  ) => StartupViewModel();
 
   @override
-  void onModelReady(StartupViewModel model) => SchedulerBinding.instance
-      .addPostFrameCallback((timeStamp) => model.runStartupLogic());
+  void onViewModelReady(StartupViewModel viewModel) => SchedulerBinding.instance
+      .addPostFrameCallback((timeStamp) => viewModel.runStartupLogic());
 }
 
 ''';
 
 // --------------------------------------------------
 
+
 // -------- App Template Data ----------
 
-const String kAppTemplateAppPath = 'lib/app/app.dart.stk';
+const String kAppTemplateAppPath =
+    'lib/app/app.dart.stk';
 
 const String kAppTemplateAppContent = '''
 import 'package:stacked/stacked_annotations.dart';
@@ -1014,9 +1052,11 @@ class App {}
 
 // --------------------------------------------------
 
+
 // -------- DialogType Template Data ----------
 
-const String kAppTemplateDialogTypePath = 'lib/enums/dialog_type.dart.stk';
+const String kAppTemplateDialogTypePath =
+    'lib/enums/dialog_type.dart.stk';
 
 const String kAppTemplateDialogTypeContent = '''
 enum DialogType {
@@ -1026,6 +1066,7 @@ enum DialogType {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- BottomSheetType Template Data ----------
 
@@ -1041,9 +1082,11 @@ enum BottomSheetType {
 
 // --------------------------------------------------
 
+
 // -------- PubspecYamlStk Template Data ----------
 
-const String kAppTemplatePubspecYamlStkPath = 'pubspec.yaml.stk';
+const String kAppTemplatePubspecYamlStkPath =
+    'pubspec.yaml.stk';
 
 const String kAppTemplatePubspecYamlStkContent = '''
 name: {{packageName}}
@@ -1147,6 +1190,7 @@ flutter:
 
 // --------------------------------------------------
 
+
 // -------- GenericViewmodelTest Template Data ----------
 
 const String kViewTemplateGenericViewmodelTestPath =
@@ -1169,6 +1213,7 @@ void main() {
 
 // --------------------------------------------------
 
+
 // -------- GenericViewmodel Template Data ----------
 
 const String kViewTemplateGenericViewmodelPath =
@@ -1182,6 +1227,7 @@ class {{viewModelName}} extends BaseViewModel {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- GenericView Template Data ----------
 
@@ -1198,7 +1244,11 @@ class {{viewName}} extends StackedView<{{viewModelName}}> {
   const {{viewName}}({Key? key}) : super(key: key);
 
   @override
-  Widget builder(BuildContext context, {{viewModelName}} model, Widget? child) {
+  Widget builder(
+    BuildContext context,
+    {{viewModelName}} viewModel,
+    Widget? child,
+  ) {
     return Scaffold(
       backgroundColor: Theme.of(context).backgroundColor,
       body: Container(
@@ -1208,13 +1258,14 @@ class {{viewName}} extends StackedView<{{viewModelName}}> {
   }
 
   @override
-  {{viewModelName}} modelBuilder(BuildContext context) {
-    return {{viewModelName}}();
-  }
+  {{viewModelName}} viewModelBuilder(
+    BuildContext context,
+  ) => {{viewModelName}}();
 }
 ''';
 
 // --------------------------------------------------
+
 
 // -------- GenericViewV1 Template Data ----------
 
@@ -1247,6 +1298,7 @@ class {{viewName}} extends StatelessWidget {
 
 // --------------------------------------------------
 
+
 // -------- GenericServiceTest Template Data ----------
 
 const String kServiceTemplateGenericServiceTestPath =
@@ -1268,6 +1320,7 @@ void main() {
 ''';
 
 // --------------------------------------------------
+
 
 // -------- GenericService Template Data ----------